move math processors to proc/math
