+++ proc/math/abs/abs.go
