+++ proc/math/mul/mulk.go
