+++ proc/math/mul/mulxy.go
