+++ proc/math/neg/negate.go
