+++ proc/math/pow/pow.go
