+++ proc/math/pow/pow_test.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/jamestunnell/go-synth/gen/array"
 	"github.com/jamestunnell/go-synth/node"
-	"github.com/jamestunnell/go-synth/proc/pow"
+	"github.com/jamestunnell/go-synth/proc/math/pow"
 	"github.com/stretchr/testify/assert"
 )
