+++ proc/math/sub/subk.go
