+++ proc/math/sub/subxy.go
